Forms create, UserController done

+++ src/main/java/web/dao/UserDaoImpl.java
@@ -2,6 +2,8 @@ package web.dao;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.transaction.Transactional;
+
 import org.springframework.stereotype.Repository;
 import web.model.User;
 
@@ -14,27 +16,35 @@ public class UserDaoImpl implements UserDao {
     EntityManager entityManager;
 
     @Override
+    @Transactional
     public List<User> listUsers() {
         return entityManager.createQuery("SELECT u FROM User u", User.class).getResultList();
     }
 
     @Override
+    @Transactional
     public void add(User user) {
         entityManager.persist(user);
     }
 
+    @Transactional
     @Override
     public void edit(User user) {
         entityManager.merge(user);
     }
 
+    @Transactional
     @Override
-    public void delete(User user) {
-        entityManager.remove(user);
+    public void delete(Long id) {
+        User user = entityManager.find(User.class, id);
+        if (user != null) {
+            entityManager.remove(user);
+        }
     }
 
+    @Transactional
     @Override
-    public User findById(int id) {
+    public User findById(Long id) {
         return entityManager.find(User.class, id);
 
     }
